refactor: change the project structure and add typing to components
