+++ src/App.tsx
@@ -2,8 +2,6 @@ import { Outlet } from 'react-router-dom';
 
 import { Header } from './components';
 
-import styles from './App.module.scss';
-
 function App() {
   return (
     <>
