+++ src/components/header/components/index.ts
@@ -1,0 +1,2 @@
+export { PopularCoin } from './popular-coin';
+export { ModalShopping } from './modal';
