+++ src/components/header/components/modal/table/cart-table.module.scss
@@ -1,0 +1,29 @@
+@use '../../../../../styles/mixins';
+
+.tableContainer {
+  width: 85%;
+  margin: 0 auto;
+}
+
+.tableField {
+  text-align: center;
+}
+
+.columnTitle {
+  color: var(--color-5);
+  font-size: var(--font-size-1);
+  font-family: var(--font-family-0);
+  text-align: center;
+}
+
+.tableWeight {
+  font-weight: var(--font-weight-3);
+}
+
+.tableIcon {
+  color: var(--color-2);
+  font-size: 3rem;
+  cursor: pointer;
+
+  @include mixins.display-center;
+}
