--- src/components/header/modal/modal.module.scss
@@ -1,29 +1,0 @@
-@use '../../../styles/mixins';
-
-.modalTable {
-
-  @include mixins.display-column(3rem,  center);
-
-  h2.modalTitle {
-    @include mixins.font-header(var(--font-size-4), (--font-weight-2), var(--color-5));
-
-  }
-}
-
-.shoppingCart {
-  @include mixins.display-column(1rem,  center);
-
-  width: 75%;
-
-  .summary {
-    @include mixins.display-row(center, center, 1rem);
-
-    p {
-      @include mixins.font-text;
-    }
-
-    span {
-      @include mixins.font-header(var(--font-size-2), var(--font-weight-3), var(--color-3))
-    }
-  }
-}
