--- src/components/header/popular-coin/popular-coin.tsx
@@ -1,14 +1,0 @@
-import styles from './popular-coin.module.scss';
-
-type PopularCoinProps = {
-  coinName: string, coinPrice: number
-}
-
-export const PopularCoin: React.FC<PopularCoinProps> = ({ coinName, coinPrice }) => {
-  return (
-    <div className={styles.coinWrapper}>
-      <h4>{coinName}</h4>
-      <p>{coinPrice}<span>$</span></p>
-    </div>
-  )
-}
