--- src/pages/info-page/coin-widget/coin-widget.module.scss
@@ -1,27 +1,0 @@
-@use '../../../styles/mixins';
-
-.widgetWrapper {
-  @include mixins.display-row(center, center, 2rem);
-
-  h1 {
-    color: var(--color-3);
-    font-family: var(--font-family-0);
-    line-height: var(--line-height);
-    font-size: var(--font-size-4);
-    font-weight: var(--font-weight-2);
-  }
-}
-
-.widgetContainer {
-  border: .3rem solid var(--color-3);
-  border-radius: .1rem;
-  padding: .7rem;
-
-  span {
-    color: var(--color-3);
-    font-family: var(--font-family-1);
-    line-height: var(--line-height);
-    font-size: var(--font-size-4);
-    font-weight: var(--font-weight-1);
-  }
-}
