+++ src/pages/info-page/components/coin-widget/index.ts
