+++ src/pages/info-page/components/index.ts
@@ -1,0 +1,4 @@
+export { CoinWidget } from './coin-widget';
+export { ItemTable } from './item-table';
+export { LineRecharts } from './line-chart';
+export { QuantityForm } from './quantity-form';
