+++ src/pages/info-page/components/item-table/index.ts
