+++ src/pages/info-page/components/item-table/item-table.module.scss
