+++ src/pages/info-page/components/line-chart/index.ts
