+++ src/pages/info-page/components/quantity-form/index.ts
