+++ src/pages/info-page/components/quantity-form/quantity-form.module.scss
@@ -1,0 +1,43 @@
+@use '../../../../styles/mixins';
+
+.formWrapper {
+
+  @include mixins.display-column(1rem, center);
+
+  width: 50%;
+  padding: 1rem;
+  background-color: var(--color-8);
+  border-radius: .1rem;
+
+  h2 {
+    font-family: var(--font-family-1);
+    font-size: var(--font-size-3);
+    line-height: var(--line-height);
+    font-weight: var(--font-weight-2);
+    color: var(--color-5);
+  }
+}
+
+.formContainer {
+  @include mixins.display-row(center, center, 1rem);
+
+  .input {
+    @include mixins.font-text;
+
+    height: 3.5rem;
+    border-radius: .1rem;
+  }
+
+  .button {
+    height: 3.3rem;
+    border-radius: .1rem;
+
+    span {
+      font-size: var(--font-size-0);
+      font-weight: var(--font-weight-2);
+      font-family: var(--font-family-1);
+      line-height: var(--line-height);
+      text-align: center;
+    }
+  }
+}
