+++ src/pages/main-page/components/index.ts
@@ -1,0 +1,2 @@
+export { TablePagination } from './pagination';
+export { DataTable } from './table';
