+++ src/pages/main-page/components/modal/index.ts
