+++ src/pages/main-page/components/modal/modal.tsx
@@ -1,0 +1,14 @@
+import { QuantityForm } from '../../../info-page/components';
+
+import styles from './modal.module.scss';
+
+export const ModalTable = () => {
+  return (
+    <div className={styles.modalTable}>
+      <h2 className={styles.modalTitle}>
+        Купить <span>Bitcoin</span>
+      </h2>
+      <QuantityForm />
+    </div>
+  );
+};
