+++ src/pages/main-page/components/pagination/index.ts
