+++ src/pages/main-page/components/pagination/pagination.module.scss
@@ -1,0 +1,7 @@
+@use '../../../../styles/mixins';
+
+.paginationContainer {
+  @include mixins.display-center;
+
+  padding: 2rem;
+}
