+++ src/pages/main-page/components/table/index.ts
